@@ -13,32 +13,32 @@
     "buildparaglide": "paraglide-js compile --project ./project.inlang --outdir ./src\\paraglide"
   },
   "devDependencies": {
-    "@iconify/json": "2.2.339",
+    "@iconify/json": "2.2.373",
     "@inlang/paraglide-js": "1.11.2",
     "@inlang/paraglide-sveltekit": "0.11.0",
-    "@playwright/test": "1.52.0",
-    "@sveltejs/adapter-auto": "3.2.5",
-    "@sveltejs/adapter-vercel": "5.7.2",
-    "@sveltejs/kit": "2.21.1",
-    "@sveltejs/vite-plugin-svelte": "3.1.2",
+    "@playwright/test": "1.54.2",
+    "@sveltejs/adapter-auto": "6.1.0",
+    "@sveltejs/adapter-vercel": "5.8.2",
+    "@sveltejs/kit": "2.30.0",
+    "@sveltejs/vite-plugin-svelte": "6.1.2",
+    "@tailwindcss/postcss": "^4.1.12",
     "@tailwindcss/typography": "^0.5.16",
     "@types/d3-scale": "4.0.9",
-    "@types/jsonwebtoken": "9.0.9",
+    "@types/jsonwebtoken": "9.0.10",
     "@types/mercadopago": "1.5.11",
     "@types/negotiator": "0.6.3",
     "@types/uuid": "9.0.8",
     "autoprefixer": "10.4.21",
     "bits-ui": "0.21.16",
-    "postcss": "8.5.3",
-    "prettier": "3.5.3",
+    "prettier": "3.6.2",
     "prettier-plugin-svelte": "3.4.0",
-    "svelte": "4.2.19",
-    "svelte-check": "3.7.1",
-    "svelte-headless-table": "0.18.2",
-    "tailwindcss": "3.4.3",
+    "svelte": "5.38.1",
+    "svelte-check": "4.3.1",
+    "svelte-headless-table": "0.18.3",
+    "tailwindcss": "4.1.12",
     "tslib": "2.8.1",
-    "typescript": "5.8.3",
-    "vite": "5.4.19",
+    "typescript": "5.9.2",
+    "vite": "7.1.2",
     "zod": "3.25.1"
   },
   "type": "module",
@@ -47,6 +47,7 @@
     "@dimforge/rapier3d-compat": "0.13.0",
     "@paypal/paypal-js": "8.2.0",
     "@tailwindcss/line-clamp": "0.4.4",
+    "@tailwindcss/vite": "^4.1.12",
     "@theatre/core": "0.7.1",
     "@theatre/studio": "0.7.1",
     "@threlte/core": "7.3.0",
@@ -63,17 +64,17 @@
     "iconify-icon": "2.1.0",
     "jsonwebtoken": "9.0.2",
     "lucide-svelte": "0.378.0",
-    "mercadopago": "2.7.0",
+    "mercadopago": "2.8.0",
     "negotiator": "0.6.3",
     "quill": "2.0.3",
     "radix-icons-svelte": "1.2.1",
     "socket.io": "4.8.1",
     "socket.io-client": "4.8.1",
-    "svelte-radix": "1.1.0",
-    "svelte-sonner": "0.3.24",
-    "tailwind-merge": "2.3.0",
-    "tailwind-variants": "0.2.1",
-    "three": "0.164.1",
+    "svelte-radix": "2.0.1",
+    "svelte-sonner": "1.0.5",
+    "tailwind-merge": "3.3.1",
+    "tailwind-variants": "2.1.0",
+    "three": "0.179.1",
     "timeago.js": "4.0.2",
     "uuid": "9.0.1"
   }
